解决warning错误

+++ async/thread.cpp
@@ -32,7 +32,7 @@ void Thread::Stop() {
 
     is_running_ = false;
     
-    for (int i = 0; i < thread_list_.size(); i++) {
+    for (size_t i = 0; i < thread_list_.size(); i++) {
         thread_list_[i]->join();
     }
 
@@ -41,7 +41,7 @@ void Thread::Stop() {
     return;
 }
 void Thread::Join() {
-    for (int i = 0; i < thread_list_.size(); i++) {
+    for (size_t i = 0; i < thread_list_.size(); i++) {
         thread_list_[i]->join();
     }
